frontend no longer computes pricing
